"Add image file type validation and improve debounce function
A new constant IMAGE_UPLOAD_FILE_TYPES was introduced in 'constants.js' to specify the accepted image file types for upload. This constant is used in 'validator.js' to check if the uploaded file type is supported. If not, an error message is displayed. A 'debounce' function was also modified in 'debounce.js' to introduce a cancellation feature which ensures that only the last delayed function call is executed to improve performance. This reduces the frequency of function calls when sorting pictures by comments, default and random in 'picture.js'."